test: split runner doctor tests
